+++ web/models.py
@@ -254,7 +254,7 @@ class HpEvent(db.Model):
 
     @classmethod
     def get_events_dicts(cls, events):
-        """From an event list to a list of event's dict with maxconf calculation once only """
+        """From an event list to a list of event's dict with max_conf calculation once only """
         if not events:
             return []
 
@@ -264,8 +264,8 @@ class HpEvent(db.Model):
         return [event.get_dict(max_conf) for event in events]
 
     def get_dict(self, max_conf):
-        td = self.stop_date - self.start_date
-        duration = datetime.timedelta(days=td.days, seconds=td.seconds, microseconds=0)
+        time_delta = self.stop_date - self.start_date
+        duration = datetime.timedelta(days=time_delta.days, seconds=time_delta.seconds, microseconds=0)
         hours_str = f"{duration}"[-8:]
         days = int(duration.days)
         duration_str = f"{days:4}d {hours_str:>8}"
@@ -469,7 +469,7 @@ class Paper(db.Model):
             self.hp_events.append(hpevent)
         self.cat_in_db = True
         db.session.commit()
-        return(len(self.hp_events))
+        return len(self.hp_events)
 
     def istex_update(self):
         """From our ids, update meta information from istex api"""
@@ -497,21 +497,10 @@ class Paper(db.Model):
 
         @return: None
         """
-        self.hp_events.clear()
+        HpEvent.query.filter_by(paper_id=self.id).delete(synchronize_session=False)
         self.cat_in_db = False
         db.session.commit()
 
-    def get_events(self):
-        """
-        Return the list of events with same doi as our's
-        #TODO: DB_REFACTOR should be better to set a new relationship between HpEvent and Paper models
-        @return:  list of HpEvents
-        """
-        found_events = []
-        doi = Doi.query.filter_by(doi=self.doi).one_or_none()
-        if doi is not None:
-            found_events = HpEvent.query.filter_by(doi_id=doi.id).all()
-        return found_events
 
     @property
     def pipeline_version(self):