+++ CHANGELOG.md
@@ -25,6 +25,15 @@ Changes are kept under subsections:
 Only major refactoring improvements or functional changes should be noted.
 
 --------------------------------------------------------------------------------
+## [2.1.0] - 2025-03-18 - Fix SUTIME current year
+### Fixed
+- current year in pipeline 8.4
+### Changed
+- show papers' events number
+### New
+- page fix_bulk: shows 2024 papers
+- more cli commands: bulk run through web 
+
 ## [2.0.0] - 2025-03-14 - Refactor DB
 #### Fixed
 - tests now all run